+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-01-03 15:55:16
+
+### Changed Files
+- `src/services/hierarchy/hierarchyService.ts`
+
+### Statistics
+```
+ 1 file changed, 35 insertions(+), 32 deletions(-)
+```
+
 ## [Checkpoint] - 2026-01-03 15:40:44
 
 ### Changed Files